# Remove location store and use `node.origin` for TIR reporting