--- a/compiler/hash-ast-expand/src/attr.rs
+++ b/compiler/hash-ast-expand/src/attr.rs
@@ -14,7 +14,6 @@ use hash_storage::store::{
 use hash_target::HasTarget;
 use hash_tir::{
     args::Arg,
-    environment::stores::tir_stores,
     lits::{CharLit, FloatLit, IntLit, Lit, StrLit},
     node::{Node, NodeOrigin},
     params::ParamIndex,
@@ -133,13 +132,10 @@ impl AstExpander<'_> {
             let mut attr = Attr::new(attr_id, node.id());
             let mut is_valid = true;
 
-            // ##GeneratedOrigin(below): these are for attributes defined in the compiler,
-            // not corresponding to any source files.
-
             // We have to build the arguments to the macro invocation
             // in the form of an `Args` and then use TIR param_utils to
             // check that the arguments and parameters match...
-            let (mac_args, args_node_id) = if let Some(mac_args) = node.args.as_ref() {
+            let (mac_args, _args_node_id) = if let Some(mac_args) = node.args.as_ref() {
                 let mut args = Vec::with_capacity(mac_args.args.len());
 
                 for (index, arg) in mac_args.args.iter().enumerate() {
@@ -183,8 +179,11 @@ impl AstExpander<'_> {
                     macro_rules! lit_prim {
                         ($name:ident,$lit_name:ident, $contents:expr) => {
                             Term::from(
-                                Term::Lit(Node::create_gen(Lit::$name($lit_name::from($contents)))),
-                                NodeOrigin::Generated,
+                                Term::Lit(Node::create_at(
+                                    Lit::$name($lit_name::from($contents)),
+                                    NodeOrigin::Given(arg.value.id()),
+                                )),
+                                NodeOrigin::Given(arg.value.id()),
                             )
                         };
                     }
@@ -200,16 +199,19 @@ impl AstExpander<'_> {
                         AttrArgIdx::from(target),
                         AttrValue { origin: arg.id(), value: attr_value },
                     );
-                    args.push(Node::at(Arg { target, value }, NodeOrigin::Generated));
+                    args.push(Node::at(Arg { target, value }, NodeOrigin::Given(arg.id())));
                 }
 
-                (Node::create_at(Node::<Arg>::seq(args), NodeOrigin::Generated), mac_args.id())
+                (
+                    Node::create_at(Node::<Arg>::seq(args), NodeOrigin::Given(node.id())),
+                    mac_args.id(),
+                )
             } else {
-                (Node::create_at(Node::<Arg>::empty_seq(), NodeOrigin::Generated), node.name.id())
+                (
+                    Node::create_at(Node::<Arg>::empty_seq(), NodeOrigin::Given(node.id())),
+                    node.name.id(),
+                )
             };
-
-            // Register the location of the args as the `mac_args` node.
-            tir_stores().ast_info().args_seq().insert(args_node_id, *mac_args.value());
 
             if is_valid && let Err(param_err) =
                 self.param_utils().validate_and_reorder_args_against_params(mac_args, attr_ty.params)