--- a/compiler/hash-exhaustiveness/src/construct.rs
+++ b/compiler/hash-exhaustiveness/src/construct.rs
@@ -30,6 +30,7 @@ use hash_source::constant::InternedStr;
 use hash_storage::store::{statics::StoreId, SequenceStoreKey, Store};
 use hash_tir::{
     data::{CtorDefId, DataTy},
+    node::NodesId,
     tuples::TupleTy,
     tys::Ty,
 };