--- a/compiler/hash-semantics/src/environment/mod.rs
+++ b/compiler/hash-semantics/src/environment/mod.rs
@@ -1,3 +1,4 @@
 // @@Docs
 pub mod analysis_progress;
+pub mod ast_info;
 pub mod sem_env;