--- a/compiler/hash-semantics/src/passes/evaluation/mod.rs
+++ b/compiler/hash-semantics/src/passes/evaluation/mod.rs
@@ -9,9 +9,9 @@ use hash_source::ModuleKind;
 use hash_storage::store::statics::SequenceStoreValue;
 use hash_tir::{
     args::Arg,
-    environment::{env::AccessToEnv, stores::tir_stores},
+    environment::env::AccessToEnv,
     fns::FnCallTerm,
-    node::Node,
+    node::{Node, NodeId},
     terms::{Term, TermId},
     utils::common::dump_tir,
 };
@@ -86,7 +86,7 @@ impl<'tc> AstPass for EvaluationPass<'tc> {
         &self,
         node: ast::AstNodeRef<ast::BodyBlock>,
     ) -> crate::diagnostics::error::SemanticResult<()> {
-        let term = tir_stores().ast_info().terms().get_data_by_node(node.id()).unwrap();
+        let term = self.ast_info().terms().get_data_by_node(node.id()).unwrap();
 
         // Potentially dump the TIR and evaluate it depending on flags.
         if self.settings().semantic_settings.dump_tir {
@@ -104,7 +104,7 @@ impl<'tc> AstPass for EvaluationPass<'tc> {
         &self,
         node: ast::AstNodeRef<ast::Module>,
     ) -> crate::diagnostics::error::SemanticResult<()> {
-        let mod_def_id = tir_stores().ast_info().mod_defs().get_data_by_node(node.id()).unwrap();
+        let mod_def_id = self.ast_info().mod_defs().get_data_by_node(node.id()).unwrap();
         let main_call_term = self.find_and_construct_main_call()?;
 
         // Potentially dump the TIR and evaluate it depending on flags.