--- a/compiler/hash-tir/src/arrays.rs
+++ b/compiler/hash-tir/src/arrays.rs
@@ -4,6 +4,7 @@ use std::fmt::Display;
 use hash_storage::store::{statics::StoreId, TrivialSequenceStoreKey};
 
 use crate::{
+    node::NodesId,
     pats::{PatId, PatListId, Spread},
     terms::{TermId, TermListId},
 };