… compiler/hash-semantics/src/passes/inference/mod.rs
… compiler/hash-semantics/src/passes/resolution/defs.rs
… compiler/hash-semantics/src/passes/resolution/exprs.rs
… compiler/hash-semantics/src/passes/resolution/mod.rs
… compiler/hash-semantics/src/passes/resolution/params.rs
… compiler/hash-semantics/src/passes/resolution/paths.rs
… compiler/hash-semantics/src/passes/resolution/pats.rs
… compiler/hash-semantics/src/passes/resolution/scoping.rs
… compiler/hash-semantics/src/passes/resolution/tys.rs
… compiler/hash-tir/src/args.rs
… compiler/hash-tir/src/atom_info.rs
… compiler/hash-tir/src/data.rs
… compiler/hash-tir/src/environment/stores.rs
… compiler/hash-tir/src/fns.rs
… compiler/hash-tir/src/locations.rs
… compiler/hash-tir/src/node.rs
… compiler/hash-tir/src/params.rs
… compiler/hash-tir/src/pats.rs
… compiler/hash-tir/src/terms.rs
… compiler/hash-tir/src/tys.rs
… compiler/hash-tir/src/utils/common.rs
… compiler/hash-tir/src/utils/mod.rs
… compiler/hash-tir/src/utils/mods.rs
… compiler/hash-tir/src/utils/params.rs
… compiler/hash-tir/src/utils/traversing.rs
… compiler/hash-typecheck/src/errors.rs
… compiler/hash-typecheck/src/inference.rs
… compiler/hash-typecheck/src/normalisation.rs
… compiler/hash-typecheck/src/substitution.rs
… compiler/hash-typecheck/src/unification.rs