Closes #942 

## Speedup metrics

The speedup after this change is about 2x improvement in typechecking!

### `printing.hash` (100k lines of hello world)

Before:
```
$ ../../target/release/hashc -i printing.hash -o out --timings

info: compiler pipeline timings:
parse             : 60.501834ms
parse::tokenise   : 31.429125ms
parse::gen        : 26.518ms
expand            : 3.586375ms
untyped-analysis  : 1.90825ms
analysis          : 2.081977959s
lower             : 72.650458ms
lower::discover   : 5.709333ms
lower::build      : 59.272542ms
lower::optimise   : 7.663125ms
codegen           : 788.437167ms
codegen::predefine: 36.417µs
codegen::build    : 64.897333ms
codegen::optimise : 33.375µs
codegen::write    : 716.586458ms
link              : 35.280625ms
link::find        : 417ns
link::execute     : 35.250292ms
build             : 3.044342668s
```

After:
```
$ ../../target/release/hashc -i printing.hash -o out --timings

info: compiler pipeline timings:
parse             : 62.680751ms
parse::tokenise   : 33.795126ms
parse::gen        : 26.456333ms
expand            : 3.596499ms
untyped-analysis  : 1.8325ms
analysis          : 1.112643041s
lower             : 52.383916ms
lower::discover   : 5.672291ms
lower::build      : 39.689875ms
lower::optimise   : 7.016292ms
codegen           : 754.809458ms
codegen::predefine: 39.917µs
codegen::build    : 62.570708ms
codegen::optimise : 30.833µs
codegen::write    : 684.675333ms
link              : 33.822167ms
link::find        : 542ns
link::execute     : 33.789917ms
build             : 2.021768332s
```